requirements.txt
python-telegram-bot==21.11.1
httpx==0.28.1
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-community==0.2.16
faiss-cpu==1.12.0
Pillow==12.1.1
pypdf==4.3.1
python-docx==1.2.0
python-pptx==1.0.2
python-dotenv==1.0.1
aiosqlite==0.20.0
